feat(forms) : added API for all forms
also
    made changes to riwayat
    hide wa, slip gaji

feat #11

+++ lib/izin/izin_list/application/jenis_izin.dart
@@ -9,7 +9,7 @@ abstract class JenisIzin with _$JenisIzin {
     // ignore: invalid_annotation_target
     @JsonKey(name: 'id_mst_izin') int? idMstIzin,
     String? nama,
-    int? qty,
+    double? qty,
     String? tipe,
   }) = _JenisIzin;
